requirements.txt
flask
flask-cors
langchain-community
faiss-cpu
openai
